move cairo0-related code to the separate module
