--- crates/cheatnet/src/execution/calls/mod.rs
@@ -1,4 +1,0 @@
-#[allow(clippy::module_name_repetitions)]
-pub mod cairo0_calls;
-#[allow(clippy::module_name_repetitions)]
-pub mod cairo1_calls;
