+++ crates/cheatnet/src/execution/deprecated/mod.rs
@@ -1,0 +1,4 @@
+pub mod cairo0_execution;
+#[allow(clippy::module_name_repetitions)]
+pub mod calls;
+pub mod syscalls;
