+++ crates/cheatnet/src/execution/entry_point.rs
@@ -1,5 +1,5 @@
 use super::cairo1_execution::execute_entry_point_call_cairo1;
-use crate::execution::cairo0_execution::execute_entry_point_call_cairo0;
+use crate::execution::deprecated::cairo0_execution::execute_entry_point_call_cairo0;
 use crate::state::CheatnetState;
 use blockifier::execution::call_info::{CallExecution, Retdata};
 use blockifier::{
