+++ crates/cheatnet/src/execution/mod.rs
@@ -1,9 +1,9 @@
 #[allow(clippy::module_name_repetitions)]
-pub mod cairo0_execution;
-#[allow(clippy::module_name_repetitions)]
 pub mod cairo1_execution;
+#[allow(clippy::module_name_repetitions)]
 pub mod calls;
 pub mod contract_print;
+pub mod deprecated;
 pub mod entry_point;
 #[allow(clippy::module_name_repetitions)]
 pub mod execution_info;
